refactors Kousa.Data.Follower -> Beef.Follows
