+++ kousa/lib/business-logic/follow_logic.ex
@@ -1,27 +1,27 @@
 defmodule Kousa.BL.Follow do
   alias Beef.Users
   alias Kousa.Gen
-  alias Kousa.Data.Follower
+  alias Beef.Follows
   alias Beef.UserBlocks
   alias Beef.Schemas.Follow
   alias Beef.Schemas.Room
 
   def get_follow_list(user_id, user_id_to_get_list_for, get_following_list, cursor) do
     if get_following_list do
-      Follower.get_following(user_id, user_id_to_get_list_for, cursor)
+      Follows.get_following(user_id, user_id_to_get_list_for, cursor)
     else
-      Follower.get_followers(user_id, user_id_to_get_list_for, cursor)
+      Follows.get_followers(user_id, user_id_to_get_list_for, cursor)
     end
   end
 
   def follow(user_id, user_you_want_to_follow_id, should_follow) do
     if should_follow do
       if user_id != user_you_want_to_follow_id and
-           not UserBlock.is_blocked(user_you_want_to_follow_id, user_id) do
-        Follower.insert(%{userId: user_you_want_to_follow_id, followerId: user_id})
+           not UserBlocks.is_blocked(user_you_want_to_follow_id, user_id) do
+        Follows.insert(%{userId: user_you_want_to_follow_id, followerId: user_id})
       end
     else
-      Follower.delete(
+      Follows.delete(
         user_you_want_to_follow_id,
         user_id
       )
@@ -30,7 +30,7 @@ defmodule Kousa.BL.Follow do
 
   @spec sync_notify_followers_you_created_a_room(String.t(), Room.t()) :: {:ok}
   def sync_notify_followers_you_created_a_room(user_id, room) do
-    followers_to_notify = Follower.get_followers_online_and_not_in_a_room(user_id)
+    followers_to_notify = Follows.get_followers_online_and_not_in_a_room(user_id)
 
     if length(followers_to_notify) > 0 do
       user = Users.get_by_id(user_id)
