+++ kousa/lib/business-logic/user_block_logic.ex
@@ -1,9 +1,9 @@
 defmodule Kousa.BL.UserBlock do
-
   alias Beef.UserBlocks
+  alias Beef.Follows
 
   def block(user_id, user_id_to_block) do
-    Kousa.Data.Follower.delete(user_id, user_id_to_block)
+    Follows.delete(user_id, user_id_to_block)
     UserBlocks.insert(%{userId: user_id, userIdBlocked: user_id_to_block})
   end
 end
